Remove unused code and restructure
