--- pkg/infra/utils.go
@@ -1,352 +1,0 @@
-package infra
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/golang/glog"
-	e2e "github.com/openshift/cluster-api-actuator-pkg/pkg/framework"
-	mapiv1beta1 "github.com/openshift/cluster-api/pkg/apis/machine/v1beta1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/client-go/discovery"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/scale"
-	"k8s.io/utils/pointer"
-	runtimeclient "sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
-	"sigs.k8s.io/controller-runtime/pkg/client/config"
-)
-
-const (
-	machineRoleLabel = "machine.openshift.io/cluster-api-machine-role"
-	machineAPIGroup  = "machine.openshift.io"
-)
-
-// machineSetsSnapShotLogs logs the state of all the machineSets in the cluster
-func machineSetsSnapShotLogs(client runtimeclient.Client) error {
-	machineSets, err := e2e.GetMachineSets(client)
-	if err != nil {
-		return fmt.Errorf("error getting machines: %v", err)
-	}
-
-	for _, machineset := range machineSets {
-		glog.Infof("MachineSet %q replicas %d. Ready: %d, available %d",
-			machineset.Name,
-			pointer.Int32PtrDerefOr(machineset.Spec.Replicas, e2e.DefaultMachineSetReplicas),
-			machineset.Status.ReadyReplicas,
-			machineset.Status.AvailableReplicas)
-	}
-	return nil
-}
-
-// getMachinesFromMachineSet returns an array of machines owned by a given machineSet
-func getMachinesFromMachineSet(client runtimeclient.Client, machineSet *mapiv1beta1.MachineSet) ([]*mapiv1beta1.Machine, error) {
-	machines, err := e2e.GetMachines(client)
-	if err != nil {
-		return nil, fmt.Errorf("error getting machines: %v", err)
-	}
-	var machinesForSet []*mapiv1beta1.Machine
-	for key := range machines {
-		if metav1.IsControlledBy(machines[key], machineSet) {
-			machinesForSet = append(machinesForSet, machines[key])
-		}
-	}
-	return machinesForSet, nil
-}
-
-// deleteMachine deletes a specific machine and returns an error otherwise
-func deleteMachine(client runtimeclient.Client, machine *mapiv1beta1.Machine) error {
-	return wait.PollImmediate(1*time.Second, time.Minute, func() (bool, error) {
-		if err := client.Delete(context.TODO(), machine); err != nil {
-			glog.Errorf("Error querying api for machine object %q: %v, retrying...", machine.Name, err)
-			return false, err
-		}
-		return true, nil
-	})
-}
-
-// GetNodesFromMachineSet returns an array of nodes backed by machines owned by a given machineSet
-func GetNodesFromMachineSet(client runtimeclient.Client, machineSet *mapiv1beta1.MachineSet) ([]*corev1.Node, error) {
-	machines, err := getMachinesFromMachineSet(client, machineSet)
-	if err != nil {
-		return nil, fmt.Errorf("error calling getMachinesFromMachineSet %v", err)
-	}
-
-	var nodes []*corev1.Node
-	for key := range machines {
-		node, err := getNodeFromMachine(client, machines[key])
-		if err != nil {
-			return nil, fmt.Errorf("error getting node from machine %q: %v", machines[key].Name, err)
-		}
-		nodes = append(nodes, node)
-	}
-	glog.Infof("MachineSet %q have %d nodes", machineSet.Name, len(nodes))
-	return nodes, nil
-}
-
-// getNodeFromMachine returns the node object referenced by machine.Status.NodeRef
-func getNodeFromMachine(client runtimeclient.Client, machine *mapiv1beta1.Machine) (*corev1.Node, error) {
-	var node corev1.Node
-	if machine.Status.NodeRef == nil {
-		glog.Errorf("Machine %q has no NodeRef", machine.Name)
-		return nil, fmt.Errorf("machine %q has no NodeRef", machine.Name)
-	}
-	key := runtimeclient.ObjectKey{Namespace: machine.Status.NodeRef.Namespace, Name: machine.Status.NodeRef.Name}
-	if err := client.Get(context.Background(), key, &node); err != nil {
-		return nil, fmt.Errorf("error getting node %q: %v", node.Name, err)
-	}
-
-	glog.Infof("Machine %q is backing node %q", machine.Name, node.Name)
-	return &node, nil
-}
-
-// NodesAreReady returns true if an array of nodes are all ready
-func NodesAreReady(nodes []*corev1.Node) bool {
-	// All nodes needs to be ready
-	for key := range nodes {
-		if !e2e.IsNodeReady(nodes[key]) {
-			glog.Errorf("Node %q is not ready. Conditions are: %v", nodes[key].Name, nodes[key].Status.Conditions)
-			return false
-		}
-		glog.Infof("Node %q is ready. Conditions are: %v", nodes[key].Name, nodes[key].Status.Conditions)
-	}
-	return true
-}
-
-// scaleMachineSet scales a machineSet with a given name to the given number of replicas
-func scaleMachineSet(name string, replicas int) error {
-	scaleClient, err := getScaleClient()
-	if err != nil {
-		return fmt.Errorf("error calling getScaleClient %v", err)
-	}
-
-	scale, err := scaleClient.Scales(e2e.MachineAPINamespace).Get(schema.GroupResource{Group: machineAPIGroup, Resource: "MachineSet"}, name)
-	if err != nil {
-		return fmt.Errorf("error calling scaleClient.Scales get: %v", err)
-	}
-
-	scaleUpdate := scale.DeepCopy()
-	scaleUpdate.Spec.Replicas = int32(replicas)
-	_, err = scaleClient.Scales(e2e.MachineAPINamespace).Update(schema.GroupResource{Group: machineAPIGroup, Resource: "MachineSet"}, scaleUpdate)
-	if err != nil {
-		return fmt.Errorf("error calling scaleClient.Scales update: %v", err)
-	}
-	return nil
-}
-
-// getScaleClient returns a ScalesGetter object to manipulate scale subresources
-func getScaleClient() (scale.ScalesGetter, error) {
-	cfg, err := config.GetConfig()
-	if err != nil {
-		return nil, fmt.Errorf("error getting config %v", err)
-	}
-	mapper, err := apiutil.NewDiscoveryRESTMapper(cfg)
-	if err != nil {
-		return nil, fmt.Errorf("error calling NewDiscoveryRESTMapper %v", err)
-	}
-
-	discovery := discovery.NewDiscoveryClientForConfigOrDie(cfg)
-	scaleKindResolver := scale.NewDiscoveryScaleKindResolver(discovery)
-	scaleClient, err := scale.NewForConfig(cfg, mapper, dynamic.LegacyAPIPathResolverFunc, scaleKindResolver)
-	if err != nil {
-		return nil, fmt.Errorf("error calling building scale client %v", err)
-	}
-	return scaleClient, nil
-}
-
-// nodesSnapShotLogs logs the state of all the nodes in the cluster
-func nodesSnapShotLogs(client runtimeclient.Client) error {
-	nodes, err := e2e.GetNodes(client)
-	if err != nil {
-		return fmt.Errorf("error getting nodes: %v", err)
-	}
-
-	for key, node := range nodes {
-		glog.Infof("Node %q. Ready: %t. Unschedulable: %t", node.Name, e2e.IsNodeReady(&nodes[key]), node.Spec.Unschedulable)
-	}
-	return nil
-}
-
-// waitUntilAllNodesAreSchedulable waits for all cluster nodes to be schedulable and returns an error otherwise
-func waitUntilAllNodesAreSchedulable(client runtimeclient.Client) error {
-	endTime := time.Now().Add(time.Duration(time.Minute))
-	return wait.PollImmediate(1*time.Second, time.Minute, func() (bool, error) {
-		nodeList := corev1.NodeList{}
-		if err := client.List(context.TODO(), &nodeList); err != nil {
-			glog.Errorf("error querying api for nodeList object: %v, retrying...", err)
-			return false, nil
-		}
-		// All nodes needs to be schedulable
-		for _, node := range nodeList.Items {
-			if node.Spec.Unschedulable == true {
-				glog.Errorf("Node %q is unschedulable", node.Name)
-				return false, nil
-			}
-			glog.Infof("[remaining %s] Node %q is schedulable", remainingTime(endTime), node.Name)
-		}
-		return true, nil
-	})
-}
-
-func waitUntilAllRCPodsAreReady(client runtimeclient.Client, rc *corev1.ReplicationController) error {
-	endTime := time.Now().Add(time.Duration(e2e.WaitLong))
-	err := wait.PollImmediate(e2e.RetryMedium, e2e.WaitLong, func() (bool, error) {
-		rcObj := corev1.ReplicationController{}
-		key := types.NamespacedName{
-			Namespace: rc.Namespace,
-			Name:      rc.Name,
-		}
-		if err := client.Get(context.TODO(), key, &rcObj); err != nil {
-			glog.Errorf("Error querying api RC %q object: %v, retrying...", rc.Name, err)
-			return false, nil
-		}
-		if rcObj.Status.ReadyReplicas == 0 {
-			glog.Infof("[%s remaining] Waiting for at least one RC ready replica, ReadyReplicas: %v, Replicas: %v", remainingTime(endTime), rcObj.Status.ReadyReplicas, rcObj.Status.Replicas)
-			return false, nil
-		}
-		glog.Infof("[%s remaining] Waiting for RC ready replicas, ReadyReplicas: %v, Replicas: %v", remainingTime(endTime), rcObj.Status.ReadyReplicas, rcObj.Status.Replicas)
-		return rcObj.Status.Replicas == rcObj.Status.ReadyReplicas, nil
-	})
-
-	// Sometimes this will timeout because Status.Replicas !=
-	// Status.ReadyReplicas. Print the state of all the pods for
-	// debugging purposes so we can distinguish between the cases
-	// when it works and those rare cases when it doesn't.
-	pods := corev1.PodList{}
-	if err := client.List(context.TODO(), &pods, runtimeclient.MatchingLabels(rc.Spec.Selector)); err != nil {
-		glog.Errorf("Error listing pods: %v", err)
-	} else {
-		prettyPrint := func(i interface{}) string {
-			s, _ := json.MarshalIndent(i, "", "  ")
-			return string(s)
-		}
-		for i := range pods.Items {
-			glog.Infof("POD #%v/%v: %s", i, len(pods.Items), prettyPrint(pods.Items[i]))
-		}
-	}
-
-	return err
-}
-
-func verifyNodeDraining(client runtimeclient.Client, targetMachine *mapiv1beta1.Machine, rc *corev1.ReplicationController) (string, error) {
-	endTime := time.Now().Add(time.Duration(e2e.WaitLong))
-	var drainedNodeName string
-	err := wait.PollImmediate(e2e.RetryMedium, e2e.WaitLong, func() (bool, error) {
-		machine := mapiv1beta1.Machine{}
-
-		key := types.NamespacedName{
-			Namespace: targetMachine.Namespace,
-			Name:      targetMachine.Name,
-		}
-		if err := client.Get(context.TODO(), key, &machine); err != nil {
-			glog.Errorf("Error querying api machine %q object: %v, retrying...", targetMachine.Name, err)
-			return false, nil
-		}
-		if machine.Status.NodeRef == nil || machine.Status.NodeRef.Kind != "Node" {
-			glog.Errorf("Machine %q not linked to a node", machine.Name)
-			return false, nil
-		}
-
-		drainedNodeName = machine.Status.NodeRef.Name
-		node := corev1.Node{}
-
-		if err := client.Get(context.TODO(), types.NamespacedName{Name: drainedNodeName}, &node); err != nil {
-			glog.Errorf("Error querying api node %q object: %v, retrying...", drainedNodeName, err)
-			return false, nil
-		}
-
-		if !node.Spec.Unschedulable {
-			glog.Errorf("Node %q is expected to be marked as unschedulable, it is not", node.Name)
-			return false, nil
-		}
-
-		glog.Infof("[remaining %s] Node %q is mark unschedulable as expected", remainingTime(endTime), node.Name)
-
-		pods := corev1.PodList{}
-		if err := client.List(context.TODO(), &pods, runtimeclient.MatchingLabels(rc.Spec.Selector)); err != nil {
-			glog.Errorf("Error querying api for Pods object: %v, retrying...", err)
-			return false, nil
-		}
-
-		podCounter := 0
-		for _, pod := range pods.Items {
-			if pod.Spec.NodeName != machine.Status.NodeRef.Name {
-				continue
-			}
-			if !pod.DeletionTimestamp.IsZero() {
-				continue
-			}
-			podCounter++
-		}
-
-		glog.Infof("[remaining %s] Have %v pods scheduled to node %q", remainingTime(endTime), podCounter, machine.Status.NodeRef.Name)
-
-		// Verify we have enough pods running as well
-		rcObj := corev1.ReplicationController{}
-		key = types.NamespacedName{
-			Namespace: rc.Namespace,
-			Name:      rc.Name,
-		}
-		if err := client.Get(context.TODO(), key, &rcObj); err != nil {
-			glog.Errorf("Error querying api RC %q object: %v, retrying...", rc.Name, err)
-			return false, nil
-		}
-
-		// The point of the test is to make sure majority of the pods are rescheduled
-		// to other nodes. Pod disruption budget makes sure at most one pod
-		// owned by the RC is not Ready. So no need to test it. Though, useful to have it printed.
-		glog.Infof("[remaining %s] RC ReadyReplicas: %v, Replicas: %v", remainingTime(endTime), rcObj.Status.ReadyReplicas, rcObj.Status.Replicas)
-
-		// This makes sure at most one replica is not ready
-		if rcObj.Status.Replicas-rcObj.Status.ReadyReplicas > 1 {
-			return false, fmt.Errorf("pod disruption budget not respected, node was not properly drained")
-		}
-
-		// Depends on timing though a machine can be deleted even before there is only
-		// one pod left on the node (that is being evicted).
-		if podCounter > 2 {
-			glog.Infof("[remaining %s] Expecting at most 2 pods to be scheduled to drained node %q, got %v", remainingTime(endTime), machine.Status.NodeRef.Name, podCounter)
-			return false, nil
-		}
-
-		glog.Infof("[remaining %s] Expected result: all pods from the RC up to last one or two got scheduled to a different node while respecting PDB", remainingTime(endTime))
-		return true, nil
-	})
-
-	return drainedNodeName, err
-}
-
-func waitUntilNodeDoesNotExists(client runtimeclient.Client, nodeName string) error {
-	endTime := time.Now().Add(time.Duration(e2e.WaitLong))
-	return wait.PollImmediate(e2e.RetryMedium, e2e.WaitLong, func() (bool, error) {
-		node := corev1.Node{}
-
-		key := types.NamespacedName{
-			Name: nodeName,
-		}
-		err := client.Get(context.TODO(), key, &node)
-		if err == nil {
-			glog.Errorf("Node %q not yet deleted", nodeName)
-			return false, nil
-		}
-
-		if !strings.Contains(err.Error(), "not found") {
-			glog.Errorf("Error querying api node %q object: %v, retrying...", nodeName, err)
-			return false, nil
-		}
-
-		glog.Infof("[%s remaining] Node %q successfully deleted", remainingTime(endTime), nodeName)
-		return true, nil
-	})
-}
-
-func remainingTime(t time.Time) time.Duration {
-	return t.Sub(time.Now()).Round(time.Second)
-}
